[ADD] Controllers were added for Comment, Price and Instructor entities. Also, some validations were updated in the request DTOs

+++ Domain/Entities/Comment.cs
@@ -9,7 +9,7 @@ namespace Domain.Entities
         public Guid CourseId { get; set; }
         public string StudentName { get; set; }
         public string Message { get; set; }
-        public int Score { get; set; }
+        public byte Score { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
         public Guid? CreatedBy { get; set; }
